Fix: modularize ux-r3 ATDD suites and mark story done
