+++ _bmad-output/implementation-artifacts/ux-r3-voicemail-and-indicator-behavior.md
@@ -1,6 +1,6 @@
 # Story ux-r3: Voicemail and Indicator Behavior
 
-Status: review
+Status: done
 
 <!-- Note: Validation is optional. Run validate-create-story for quality check before dev-story. -->
