+++ tests/support/fixtures/connectShyftStoryUxR3.fixture.ts
@@ -1,10 +1,13 @@
-import { randomUUID } from 'node:crypto';
 import { test as base } from '@playwright/test';
 import {
   createStoryUxR3Context,
   createStoryUxR3Headers,
   type StoryUxR3Context,
 } from '../factories/connectShyftStoryUxR3Factory';
+import {
+  deterministicProviderEventId,
+  deterministicToken,
+} from '../utils/deterministicTestIds';
 
 type StoryUxR3Fixtures = {
   storyUxR3Context: StoryUxR3Context;
@@ -17,8 +20,19 @@ type StoryUxR3Fixtures = {
 };
 
 export const test = base.extend<StoryUxR3Fixtures>({
-  storyUxR3Context: async ({}, use) => {
-    await use(createStoryUxR3Context());
+  storyUxR3Context: async ({}, use, testInfo) => {
+    await use(
+      createStoryUxR3Context({
+        correlationId: `corr-story-ux-r3-${deterministicToken(
+          testInfo,
+          'uxr3-fixture-correlation',
+        )}`,
+        csrfToken: `csrf-story-ux-r3-${deterministicToken(
+          testInfo,
+          'uxr3-fixture-csrf',
+        )}`,
+      }),
+    );
   },
   storyUxR3MemberHeaders: async ({ storyUxR3Context }, use) => {
     await use(
@@ -52,11 +66,18 @@ export const test = base.extend<StoryUxR3Fixtures>({
     });
     await use(`?${params.toString()}`);
   },
-  storyUxR3InboundVoicemailPayload: async ({ storyUxR3Context }, use) => {
+  storyUxR3InboundVoicemailPayload: async ({ storyUxR3Context }, use, testInfo) => {
     await use({
       provider: 'telnyx',
-      providerEventId: `evt-ux-r3-voice-${randomUUID().slice(0, 8)}`,
-      providerLegId: `leg-ux-r3-voice-${randomUUID().slice(0, 8)}`,
+      providerEventId: deterministicProviderEventId(
+        'provider-event-uxr3-fixture',
+        testInfo,
+        'uxr3-voicemail-event',
+      ),
+      providerLegId: `leg-uxr3-fixture-${deterministicToken(
+        testInfo,
+        'uxr3-voicemail-leg',
+      )}`,
       eventType: storyUxR3Context.events.inboundVoicemail,
       tenantId: storyUxR3Context.tenantId,
       orgUnitId: storyUxR3Context.orgUnitId,
@@ -67,11 +88,18 @@ export const test = base.extend<StoryUxR3Fixtures>({
       },
     });
   },
-  storyUxR3InboundClosedPayload: async ({ storyUxR3Context }, use) => {
+  storyUxR3InboundClosedPayload: async ({ storyUxR3Context }, use, testInfo) => {
     await use({
       provider: 'telnyx',
-      providerEventId: `evt-ux-r3-closed-${randomUUID().slice(0, 8)}`,
-      providerLegId: `leg-ux-r3-closed-${randomUUID().slice(0, 8)}`,
+      providerEventId: deterministicProviderEventId(
+        'provider-event-uxr3-fixture',
+        testInfo,
+        'uxr3-closed-event',
+      ),
+      providerLegId: `leg-uxr3-fixture-${deterministicToken(
+        testInfo,
+        'uxr3-closed-leg',
+      )}`,
       eventType: storyUxR3Context.events.inboundMissedCall,
       tenantId: storyUxR3Context.tenantId,
       orgUnitId: storyUxR3Context.orgUnitId,
